+++ brest/interfaces/__init__.py
@@ -1,7 +1,9 @@
 from .interfaces import Interfaces
+from .serial_interface import SerialInterface
 
 __all__ = [
     'Interfaces',
+    'SerialInterface',
 ]
 
 __version__ = '0.0.6'